Added more tests to hopefully have enough coverage
